added features discussed today. For more information see below.
I created a SqlOutputWindow and a SqlOutputWindowDAO for the database transactions.
I removed all the test buttons after showing Professor,
and he wanted the show tables button moved to the sql panel on Elle_Gui.
This now works with only one frame so if show tables button or the enter button is used, 
it will either move the open one to the front or open a new one if not open.
I set the width of the output window to the same width as Elle_Gui.
Only select, show, or describe is allowed and an appropriate message is displayed otherwise.
Cleaned up the old code and behavior because it is no longer needed.

This is the method that I use to set the table model for the table. 
I used JXTable for the packAll().
If the table width was greater than the scroll pane width,
 then I used the packAll() and turn autosizing off.
If the table width was less than the scroll pane width then I used the autosizing for all columns.

private boolean setTableModel(String sqlCommand) {
        DefaultTableModel model = sqlOutputWindowDAO.getTableModel(sqlCommand);
        if(model != null){
            tableOutput.setAutoResizeMode( JTable.AUTO_RESIZE_OFF );
            tableOutput.setModel(model);
            tableOutput.packAll();
            if(tableOutput.getPreferredSize().width < scrollPaneOutput.getWidth()){
                tableOutput.setAutoResizeMode( JTable.AUTO_RESIZE_ALL_COLUMNS );
            }
            return true;
        }
        else{
            return false;
        }

    }